+++ Platforms/VulkanPlatform/include/MyGUI_VulkanRTTexture.h
@@ -43,6 +43,7 @@ namespace MyGUI
 		VkFramebuffer mFramebuffer{VK_NULL_HANDLE};
 		VkCommandBuffer mCommandBuffer{VK_NULL_HANDLE};
 		VkFence mFence{VK_NULL_HANDLE};
+		bool mPending{false};
 		uint32_t mWidth{0};
 		uint32_t mHeight{0};
 		RenderTargetInfo mRenderTargetInfo;